+++ CHANGES.rst
@@ -18,6 +18,7 @@ Version 0.9.0 (UNRELEASED)
 - Changes the workflow-details page to show the logs of the workflow engine.
 - Changes the workflow-details page to show file sizes in a human-readable format.
 - Changes the workflow-details page to show the workspace's retention rules.
+- Changes the workflow-details page to show the duration of the workflow's jobs.
 - Changes the deletion of a workflow to always clean up the workspace.
 
 Version 0.8.2 (2022-02-15)